+++ .storybook/preview.js
@@ -2,11 +2,10 @@ import "../src/nationalarchives/all.scss";
 import { a11yConfig } from "./storybook-config";
 import { MINIMAL_VIEWPORTS } from "@storybook/addon-viewport";
 
-// Also need to change $support-colour-schemes in src/nationalarchives/_features.scss
-export const supportDynamicColourSchemes = false;
+export const dynamicColourSchemeStories = false;
 document.documentElement.classList.add(
   "tna-template",
-  supportDynamicColourSchemes
+  dynamicColourSchemeStories
     ? "tna-template--system-theme"
     : "tna-template--light-theme",
 );